+++ aries-site/src/examples/templates/feedback/ModalTaskFlowExample/index.js
@@ -1,0 +1,3 @@
+export * from './PowerDeviceFeedback';
+export * from './PowerOnBodyExample';
+export * from './SolicitedFeedbackExample';
