+++ aries-site/src/examples/templates/feedback/index.js
@@ -1,5 +1,3 @@
 export * from './FeedbackAnatomy';
-export * from './PowerDeviceFeedback';
-export * from './PowerOnBodyExample';
-export * from './SolicitedFeedbackExample';
+export * from './ModalTaskFlowExample';
 export * from './SimpleFeedbackExample';
